Make CRI logs parsing to a library
